Implement workaround for unequip bug
The engine has a bug whereby a weapon in the secondary (leftmost) slot gets moved to the primary slot when a scope is attached/detached. If the primary slot is already full, the weapon is moved from player's hands into their backpack. The addon now detects this and re-equips the weapon after scope change.

Bump version to 1.0 as I now consider this addon to be feature-complete.