Fix the error when toggle button is pressed with no compatible scope in inventory
When holding a weapon with a native scope (a single-scope weapon such as Galil), pressing the toggle key caused the addon to get stuck in an invalid state which disabled the toggle key until game reload.

Reminder: do not put a comment saying "should not happen" in a code path you *know* to be unreachable. It will happen. Repeatedly. On the customer's production site.