Default to weapon firemode when queue size not available.
When the HUD is turned off, m_fire_mode text may be blank. If we use this to "reset the firing mode", we break the weapon by giving it a queue size of zero.

If we get a blank queue size, we will default to the actual firing mode from the weapon. Unfortunately, this is out-of-sync with the queue size method that we use to "preserve the firing mode", so it will work only once. After we toggle scope a second time, the weapon will revert to full-auto.

Also add/fix some trace calls.